backend/requirements.txt
flask_cors==4.0.0
flask_jwt_extended==4.5.3
python-dotenv==1.0.0
flask==3.0.0
pymongo==4.5.0
o365==2.0.26